requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
dnspython==2.2.1
et-xmlfile==1.1.0
eventlet==0.33.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
model-bakery==1.5.0
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
ujson==5.1.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
